// src/core/models/FileAttachment.mode.ts
export class FileAttachment {
  // id?: number;
  // fileId?: number;
  path?: string;
  name?: string;
  type?: string;
  size?: number;

  file?: File | null;
  // isUploaded?: boolean | null;
  // isDelete?: boolean | null;
  // createDatetime?: string | null;
  // lastUpdateDatetime?: string | null;

  // loader?: boolean = false;
  // progress?: number = 0;
  // uploadId?: string;
}